@@ -2,8 +2,9 @@ cmake_minimum_required(VERSION 2.8)
 
 project(Athena)
 
+set(CMAKE_CXX_STANDARD 14)
 set(CMAKE_MODULE_PATH ${CMAKE_MODULE_PATH} "${PROJECT_SOURCE_DIR}/cmake")
-set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -std=c++11 -O3")
+set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -std=c++14 -O3")
 
 # SDL2, SDL2/SDL2_Image, SDL2/SDL2_ttf
 find_package(SDL2 REQUIRED)
@@ -20,8 +21,7 @@ file(GLOB SOURCES src/*.cpp src/*.c src/*/*.cpp src/*/*.c src/*/*/*.cpp src/*/*/
 file(GLOB HEADERS src/*.hpp src/*.h src/*/*.hpp src/*/*.h src/*/*/*.hpp src/*/*/*.h include/*/*.hpp include/*/*.h)
 
 
-
-add_executable(${PROJECT_NAME} ${SOURCES} ${HEADERS})
+add_executable(${PROJECT_NAME} ${SOURCES} ${HEADERS} src/TrECS/TrSystems/TrEntitySystem.cpp src/TrECS/TrSystems/TrEntitySystem.hpp)
 #add_executable(${PROJECT_NAME} "src/TrGame.cpp" ${SOURCES} ${HEADERS})
 target_link_libraries(${PROJECT_NAME}
         ${SDL2_LIBRARY}
