@@ -3,8 +3,9 @@
  */
 
 #include "TrFootprintComponent.hpp"
+#include "../TrEntities.hpp"
 
-void TrFootprintComponent::render(TrGame* game, TrEntity* entity) {
+void TrFootprintComponent::render(TrGame *game, TrEntity *entity) {
   for (int y = 0; y < m_h; y++) {
     for (int x = 0; x < m_w; x++) {
       if (m_cells[x + m_w * y]) {
