@@ -1,7 +1,7 @@
 #pragma once
 
 #include "../TrComponents.hpp"
-#include "../TrEntities/TrBuildingEntity.hpp"
+
 /**
  * @brief Buildings have variable footprints, this keeps track of them
  */
@@ -9,11 +9,14 @@ class TrFootprintComponent : public TrComponent {
  public:
   int m_w;
   int m_h;
-  bool* m_cells;
+  vector<char> m_cells;
 
-  TrFootprintComponent(int w, int h, bool* footprint)
-      : m_w(w), m_h(h), m_cells(footprint){};
+  TrFootprintComponent(int w, int h, vector<char> footprint)
+      : m_w(w), m_h(h), m_cells(move(footprint)) {};
 
-  void update(TrGame* game, TrEntity* entity) override {};
-  virtual void render(TrGame* game, TrEntity* entity);
+  TrFootprintComponent(int w, int h)
+      : m_w(w), m_h(h) {};
+
+  void update(TrGame *game, TrEntity *entity) override {};
+  virtual void render(TrGame *game, TrEntity *entity);
 };
