@@ -7,6 +7,7 @@
 #include <set>
 
 #include "../TrComponents.hpp"
+
 class TrEntity;
 
 using namespace std;
@@ -16,18 +17,18 @@ using namespace std;
  */
 class TrGraphicsComponent : public TrComponent {
  public:
-  SDL_Texture* m_texture;
+  SDL_Texture *m_texture;
   SDL_Color m_color;
 
-  explicit TrGraphicsComponent(SDL_Texture* texture)
+  explicit TrGraphicsComponent(SDL_Texture *texture)
       : m_texture(texture), m_color{0, 0, 0, 0} {};
-  explicit TrGraphicsComponent(SDL_Color color) : m_texture(nullptr), m_color(color){};
+  explicit TrGraphicsComponent(SDL_Color color) : m_texture(nullptr), m_color(color) {};
   TrGraphicsComponent() : TrGraphicsComponent(nullptr) {}
 
-  ~TrGraphicsComponent() override = default;;
+  ~TrGraphicsComponent() override = default;
 
   /**
    * @brief draws the given entity
    */
-  void update(TrGame* game, TrEntity* entity) override;
+  void update(TrGame *game, TrEntity *entity) override;
 };
