@@ -4,7 +4,7 @@
 
 #include "TrPhysicsComponent.hpp"
 
-void TrPhysicsComponent::update(TrGame* game, TrEntity* entity) {
+void TrPhysicsComponent::update(TrGame *game, TrEntity *entity) {
   m_vx += m_ax;
   m_vy += m_ay;
 
