@@ -5,4 +5,4 @@
 #include "TrPlanningComponent.hpp"
 
 // TODO: something
-void TrPlanningComponent::update(TrGame* game, TrEntity* entity) {}
+void TrPlanningComponent::update(TrGame *game, TrEntity *entity) {}
