@@ -11,15 +11,17 @@
 #include "TrComponents/TrPhysicsComponent.hpp"
 #include "TrComponents/TrPlanningComponent.hpp"
 
+class TrGame;
+
 /**
  * @brief General interface for all game entities
  */
 class TrEntity {
  public:
-  TrGame* m_game;
+  TrGame *m_game;
 
   SDL_Rect m_rect;
 
-  TrEntity(TrGame* game, SDL_Rect rect) : m_game(game), m_rect(rect) {}
-  virtual ~TrEntity() = 0;
+  TrEntity(TrGame *game, SDL_Rect rect) : m_game(game), m_rect(rect) {}
+  virtual ~TrEntity() {}; // TODO: make TrEntity uninstantiable
 };
