@@ -4,9 +4,9 @@
  * TrActorEntity.hpp
  */
 
-#include "../TrComponents/TrGraphicsComponent.hpp"
 #include "../TrEntities.hpp"
 
+class TrGame;
 class TrActorEntityType;
 
 /**
@@ -14,10 +14,10 @@ class TrActorEntityType;
  */
 class TrActorEntity : public TrEntity {
  public:
-  TrActorEntityType* m_type;
+  TrActorEntityType *m_type;
 
-  TrActorEntity(TrGame* game, SDL_Rect rect, TrActorEntityType* type)
-      : TrEntity(game, rect), m_type(type){};
+  TrActorEntity(TrGame *game, SDL_Rect rect, TrActorEntityType *type)
+      : TrEntity(game, rect), m_type(type) {};
 
   ~TrActorEntity() override = default;
 };
