@@ -15,9 +15,9 @@ class TrFormula {
  public:
   TrFormula() = default;
 
-  map<TrItem*, int> m_inputs;
-  map<TrItem*, int> m_outputs;
-  map<TrItem*, int> m_byproducts;
+  map<TrItem *, int> m_inputs;
+  map<TrItem *, int> m_outputs;
+  map<TrItem *, int> m_byproducts;
 
   virtual ~TrFormula() = default;
 };
