@@ -11,13 +11,15 @@
 using namespace std;
 class TrGUIDropdownMenu : public TrGUIObject {
  public:
-  TrGUIButton* m_button;
-  TrGUIMenu* m_menu;
+
+  unique_ptr<TrGUIButton> m_button;
+  unique_ptr<TrGUIMenu> m_menu;
   bool m_isMenuOpen;
 
   // TODO: how ot call parent constructor
 
-  TrGUIDropdownMenu(TrGame* game, TrGUIButton* button, TrGUIMenu* menu);
+  TrGUIDropdownMenu(TrGame *game, unique_ptr<TrGUIButton> button, unique_ptr<TrGUIMenu>
+  menu);
   ~TrGUIDropdownMenu() override;
 
   void draw() override;
