@@ -15,10 +15,10 @@ class TrGUIMenu : public TrGUIObject {
   int m_spacing = 10;
   vector<TrGUIObject *> m_buttons;
 
-  static TrGUIMenu *MakeHorizontalMenu(TrGame *game, SDL_Rect rect,
-                                       vector<string> labels);
-  static TrGUIMenu *MakeVerticalMenu(TrGame *game, SDL_Rect rect,
-                                     vector<string> labels);
+  static unique_ptr<TrGUIMenu> MakeHorizontalMenu(TrGame *game, SDL_Rect rect,
+                                                  vector<string> labels);
+  static unique_ptr<TrGUIMenu> MakeVerticalMenu(TrGame *game, SDL_Rect rect,
+                                                vector<string> labels);
 
   ~TrGUIMenu() override;
 
