@@ -1,4 +1,4 @@
 
 #include "TrWindMap.hpp"
 
-void TrWindMap::update(TrMap* map) {}
+void TrWindMap::update(TrMap *map) {}
