@@ -3,11 +3,10 @@
  */
 
 #include "TrTransitionLoop.hpp"
-#include "../TrGame.hpp"
 
 class TrMainMenuLoop;
 
-TrRenderLoop* TrTransitionLoop::update(TrGame* game) {
+TrRenderLoop *TrTransitionLoop::update(TrGame *game) {
   if (m_waitTick && m_waitTick++ < m_maxWaitTick) {
     return this;
   }
@@ -25,7 +24,7 @@ TrRenderLoop* TrTransitionLoop::update(TrGame* game) {
   return nullptr;
 }
 
-void TrTransitionLoop::render(TrGame* game) {
+void TrTransitionLoop::render(TrGame *game) {
   int mx;
   int my;
 
@@ -36,37 +35,35 @@ void TrTransitionLoop::render(TrGame* game) {
   // getting coarser
   if (m_curTick < m_maxTicks / 2) {
     mx = 20 +
-         (K_MAP_SIZE_X - 20) * ((float)((m_maxTicks / 2.0) - m_curTick) /
-                                (float)((m_maxTicks / 2.0)));
+        (K_MAP_SIZE_X - 20) * ((float) ((m_maxTicks / 2.0) - m_curTick) /
+            (float) ((m_maxTicks / 2.0)));
     my = 20 +
-         (K_MAP_SIZE_Y - 20) * ((float)((m_maxTicks / 2.0) - m_curTick) /
-                                (float)((m_maxTicks / 2.0)));
+        (K_MAP_SIZE_Y - 20) * ((float) ((m_maxTicks / 2.0) - m_curTick) /
+            (float) ((m_maxTicks / 2.0)));
   } else {
     // getting finer
     mx = 20 +
-         (K_MAP_SIZE_X - 20) * ((float)(m_curTick - (m_maxTicks / 2.0)) /
-                                (float)((m_maxTicks / 2.0)));
+        (K_MAP_SIZE_X - 20) * ((float) (m_curTick - (m_maxTicks / 2.0)) /
+            (float) ((m_maxTicks / 2.0)));
     my = 20 +
-         (K_MAP_SIZE_Y - 20) * ((float)(m_curTick - (m_maxTicks / 2.0)) /
-                                (float)((m_maxTicks / 2.0)));
+        (K_MAP_SIZE_Y - 20) * ((float) (m_curTick - (m_maxTicks / 2.0)) /
+            (float) ((m_maxTicks / 2.0)));
   }
 
-  // TODO: is this slow?
-  SDL_Texture* tempTexSmall =
+  SDL_Texture *tempTexSmall =
       SDL_CreateTexture(game->m_SDLRenderer, SDL_PIXELFORMAT_RGBA8888,
                         SDL_TEXTUREACCESS_TARGET, mx, my);
-  SDL_Texture* tempTexFull = SDL_CreateTexture(
+  SDL_Texture *tempTexFull = SDL_CreateTexture(
       game->m_SDLRenderer, SDL_PIXELFORMAT_RGBA8888, SDL_TEXTUREACCESS_TARGET,
       sz(K_DISPLAY_SIZE_X), sz(K_DISPLAY_SIZE_Y));
 
   SDL_SetTextureBlendMode(tempTexSmall, SDL_BLENDMODE_BLEND);
   SDL_SetTextureBlendMode(tempTexFull, SDL_BLENDMODE_BLEND);
 
-  // TODO: make sure mapsize - m_curtick * whatev isnt negative
   SDL_Rect tmp = {0, 0, mx, my};
   SDL_Rect tmpFr = {0, 0, mx, my};
-  SDL_Rect screenQuad = {0, 0, sz((K_MAP_SIZE_X)*K_DISPLAY_SCALE),
-                         sz((K_MAP_SIZE_Y)*K_DISPLAY_SCALE)};
+  SDL_Rect screenQuad = {0, 0, sz((K_MAP_SIZE_X) * K_DISPLAY_SCALE),
+                         sz((K_MAP_SIZE_Y) * K_DISPLAY_SCALE)};
 
   SDL_SetRenderTarget(game->m_SDLRenderer, tempTexFull);
 
@@ -87,14 +84,14 @@ void TrTransitionLoop::render(TrGame* game) {
     alpha = m_maxTicks - m_curTick;
   }
 
-  alpha = alpha / (float)(m_maxTicks / 2.0);
+  alpha = alpha / (float) (m_maxTicks / 2.0);
 
   alpha = sqrt(1.0 - alpha) * 0.9 + 0.1;
 
   alpha *= 255.0;
 
   SDL_SetRenderTarget(game->m_SDLRenderer, nullptr);
-  SDL_SetTextureAlphaMod(tempTexSmall, (int)alpha);
+  SDL_SetTextureAlphaMod(tempTexSmall, (int) alpha);
 
   SDL_RenderCopy(game->m_SDLRenderer, tempTexSmall, &tmpFr, &screenQuad);
 
