@@ -4,11 +4,11 @@
  * TrQuadTree.hpp
  */
 
-template <class T>
+template<class T>
 class TrQuadTree {
  public:
   T m_dat;
-  TrQuadTree* m_children[4];
+  TrQuadTree *m_children[4];
 
   TrQuadTree();
   virtual ~TrQuadTree();
