@@ -35,7 +35,7 @@
 using namespace std;
 using namespace glm;
 
-void renderTextureWithOffset(SDL_Renderer* renderer, SDL_Texture* texture,
+void renderTextureWithOffset(SDL_Renderer *renderer, SDL_Texture *texture,
                              int xOff, int yOff, int pixelSize);
 
 // edited from
@@ -55,43 +55,35 @@ inline double lerp3(double omin, double omax, double xx) {
 }
 
 inline dvec3 colorToVec(uint32_t color) {
-  //   ocolor.x = (this->at(i, j) & 0x00FF0000) >> 16;
-  // ocolor.y = (this->at(i, j) & 0x0000FF00) >> 8;
-  // ocolor.z = this->at(i, j) & 0x000000FF;
-
   return dvec3((color & 0x00FF0000u) >> 16u, (color & 0x0000FF00u) >> 8u,
                color & 0x000000FFu);
 }
 
 inline uint32_t vecToColor(dvec3 color) {
-  return 0xFF000000u | ((uint)color.x & 0xFFu) << 16u | ((uint)color.y & 0xFFu) << 8u |
-         ((uint)color.z & 0xFFu);
+  return 0xFF000000u | ((uint) color.x & 0xFFu) << 16u | ((uint) color.y & 0xFFu) << 8u |
+      ((uint) color.z & 0xFFu);
 }
 
 inline double clockToMilliseconds(clock_t ticks) {
-  return (ticks / (double)CLOCKS_PER_SEC) * 1000.0;
+  return (ticks / (double) CLOCKS_PER_SEC) * 1000.0;
 }
 
 inline uint32_t lerpColor(uint32_t color1, uint32_t color2, double lerp) {
-  int cr1 = (color1 & K_R_MASK) >> 16u;
-  int cg1 = (color1 & K_G_MASK) >> 8u;
-  int cb1 = color1 & K_B_MASK;
+  uint32_t cr1 = (color1 & K_R_MASK) >> 16u;
+  uint32_t cg1 = (color1 & K_G_MASK) >> 8u;
+  uint32_t cb1 = color1 & K_B_MASK;
 
-  int cr2 = (color2 & K_R_MASK) >> 16u;
-  int cg2 = (color2 & K_G_MASK) >> 8u;
-  int cb2 = color2 & K_B_MASK;
+  uint32_t cr2 = (color2 & K_R_MASK) >> 16u;
+  uint32_t cg2 = (color2 & K_G_MASK) >> 8u;
+  uint32_t cb2 = color2 & K_B_MASK;
 
-  int cr = cr1 + (int)round((cr2 - cr1) * lerp);
-  int cg = cg1 + (int)round((cg2 - cg1) * lerp);
-  int cb = cb1 + (int)round((cb2 - cb1) * lerp);
+  uint32_t cr = cr1 + (uint32_t) round((cr2 - cr1) * lerp);
+  uint32_t cg = cg1 + (uint32_t) round((cg2 - cg1) * lerp);
+  uint32_t cb = cb1 + (uint32_t) round((cb2 - cb1) * lerp);
 
   if (cr > 255) cr = 255;
   if (cg > 255) cg = 255;
   if (cb > 255) cb = 255;
-
-  if (cr < 0) cr = 0;
-  if (cg < 0) cg = 0;
-  if (cb < 0) cb = 0;
 
   uint32_t ret = K_A_MASK;
   ret |= cr << 16u;
@@ -102,21 +94,17 @@ inline uint32_t lerpColor(uint32_t color1, uint32_t color2, double lerp) {
 }
 
 inline uint32_t multiplyColor(uint32_t color, double r, double g, double b) {
-  int cr = (color & K_R_MASK) >> 16u;
-  int cg = (color & K_G_MASK) >> 8u;
-  int cb = color & K_B_MASK;
+  uint32_t cr = (color & K_R_MASK) >> 16u;
+  uint32_t cg = (color & K_G_MASK) >> 8u;
+  uint32_t cb = color & K_B_MASK;
 
-  cr = floor(cr * r);
-  cg = floor(cg * g);
-  cb = floor(cb * b);
+  cr = static_cast<uint32_t>(floor(cr * (r > 0 ? r : 0)));
+  cg = static_cast<uint32_t>(floor(cg * (g > 0 ? g : 0)));
+  cb = static_cast<uint32_t>(floor(cb * (b > 0 ? b : 0)));
 
   if (cr > 255) cr = 255;
   if (cg > 255) cg = 255;
   if (cb > 255) cb = 255;
-
-  if (cr < 0) cr = 0;
-  if (cg < 0) cg = 0;
-  if (cb < 0) cb = 0;
 
   uint32_t ret = K_A_MASK;
   ret |= cr << 16u;
